+++ src/python/DAS/services/abstract_service.py
@@ -600,7 +600,7 @@ class DASAbstractService(object):
         for further processing.
         """
         cond  = getarg(query, 'spec', {})
-        instance = cond.get('instance', '')
+        instance = query.get('instance', 'cms_dbs_prod_global')
         skeys = getarg(query, 'fields', [])
         if  not skeys:
             skeys = []