+++ doc/sphinx/changelog.rst
@@ -9,6 +9,9 @@ performed stress tests and code audit DAS servers.
 
 - 0.6.3
 
+  - put instance as separate key into mongo query (it's ignored everywhere except DBS)
+  - work on web UI (remove view code/yaml), put dbs instances, remember
+    user settings for view/instance on a page
   - add physics group to DBS2 queries
   - add support to look-up of sites for a given dataset/block
   - allow to use pattern in filters, e.g. block.replica.site=*T1*